+++ config/prod.ts
@@ -1,0 +1,13 @@
+const prod: CFHostConfig = {
+  sentryKey: 'https://c2a21df7b3794fc6a8854f0a51e0e790@sentry.choiceform.io/8',
+  recordUrl: 'https://tracking.choiceform.com/api/records',
+  host: 'https://osapi.choiceform.com',
+  wsHost: 'wss://osapi.choiceform.com',
+  cdnHost: 'https://media.choiceform.com',
+  oldIconHost: 'https://cform-files.b0.upaiyun.com/UploadFiles/UploadSvg',
+  publicHost: 'https://public.choiceform.com',
+  dictUrl:
+    'https://media.choiceform.com/os-editor/assets/UploadFiles/Dictionary/Dictionary.js',
+}
+
+export default prod;