+++ config/staging.ts
@@ -1,0 +1,13 @@
+const staging: CFHostConfig = {
+  recordUrl: 'https://tracking.choiceform.io/api/records',
+  host: 'https://osapi.choiceform.io',
+  wsHost: 'wss://osapi.choiceform.io',
+  cdnHost: 'https://media.choiceform.io',
+  oldIconHost: 'https://cform-files.b0.upaiyun.com/UploadFiles/UploadSvg',
+  publicHost: 'https://public.choiceform.io',
+  sentryKey: 'https://15f81d0f358545a586747834c943bd88@sentry.choiceform.io/7',
+  dictUrl:
+    'https://media.choiceform.com/os-editor/assets/UploadFiles/Dictionary/Dictionary.js',
+};
+
+export default staging;