+++ types/basic.d.ts
@@ -1,0 +1,4 @@
+declare module 'config' {
+  const CF_CONFIG: CFHostConfig;
+  export default CF_CONFIG;
+}